feat(middleware): add progress lifecycle
added progress lifecycle for use when a retry is passed in but the consumer would still like to get
the updated resource each time it is fetched even though the final resource is not yet ready for
whatever reason. an example usecase for this is a job that reports percent complete. we still want
to know when the job is done and thus cant piggyback on success dispatch so consumers can still
control things like loading indication.

BREAKING CHANGE: progress lifecycle will now dispatch a progress action when something is retried